requirements.txt
Django>=5.0.6,<5.1
psycopg2>=2.9.9,<2.10
boto3>=1.34.127<1.35
django-storages>=1.14.3<1.15
google-auth>=2.30.0<=2.31
google-auth-oauthlib>=1.2.0<1.3
google-auth-httplib2>=0.2.0<0.3
requests>=2.32.3<2.33
PyJWT>=2.8.0<2.9
